+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   used to toggle the enforcement of the limit.
 * New config `cider-interactive-eval-result-prefix` controls the prefix displayed before results
 from interactive evaluation displayed in the minibuffer.
+* New config `cider-repl-result-prefix` controls the prefix displayed before results in the REPL.
 * Font-lock interactive evaluation results as Clojure code.
 * Added the ability to font-lock input and results in the REPL as Clojure code. This is controlled via
 the option `cider-repl-use-clojure-font-lock`.